requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
build==1.0.3
Cerberus==1.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2
django-allauth==0.58.2
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-pdb==0.6.2
djangorestframework==3.14.0
docopt==0.6.2
drf-spectacular==0.26.5
environs==9.5.0
factory-boy==3.3.0
Faker==22.5.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.5.1
marshmallow==3.20.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.0
pep517==0.13.1
pip-api==0.0.30
pip-tools==7.3.0
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pluggy==1.3.0
psycopg==3.1.14
psycopg-binary==3.1.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
requirementslib==3.0.0
rpds-py==0.13.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomlkit==0.12.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
yarg==0.1.9
